Introduced functionality to check if a study/series/instance has changed. This is achieved by using ETags. For Study and Series. ETag is calculated using the following formula:     Max(Instance watermark)-Count(Instance) For Instance, respective watermark serves as the ETag. Calculation of ETag is done in a newly introduced stored procedure called GetETag. This calculated ETag is then appended to the list of headers in the response.
When a request is received, we check for If-None-Match header field. If this is present, we compare it to the calculated ETag which indicates whether the content has changed since the last time user had retrieved this metadata. If the content has not changed, we send back Http Status Code 304 (not modified). If the content has changed, we send back proper response with updated ETag appended to the list of headers.

Related tests are also updated.