IMS: Support a way to check if a Study/Series/Instance has changed (#205)
* - Added health check feature.
- Uri: /health/check
- Checks for health of Blobs and Sql Server. Returns unhealthy if either of these services is unhealthy.
- Sample output:
    {"overallStatus":"Healthy","details":[{"name":"DicomBlobHealthCheck","status":"Healthy","description":"Successfully connected to the blob data store."},{"name":"SqlServerHealthCheck","status":"Healthy","description":"Successfully connected to the data store."}]}
- Added tests for them as well.

* Updating NSubstitute.dll to the latest version.

* Addressed comments.

* Updated DICOM code after health check shared code was moved to healthcare shared components project.

* Fixed interfaces for the shared health check code.

* Introduced functionality to check if a study/series/instance has changed. This is achieved by using ETags.
For Study and Series. ETag is calculated using the following formula:
    Max(Instance watermark)-Count(Instance)
For Instance, respective watermark serves as the ETag.
Calculation of ETag is done in a newly introduced stored procedure called GetETag.
This calculated ETag is then appended to the list of headers in the response.

When a request is received, we check for If-None-Match header field. If this is present, we compare it to the calculated ETag which indicates whether the content has changed since the last time user had retrieved this metadata. If the content has not changed, we send back Http Status Code 304 (not modified). If the content has changed, we send back proper response with updated ETag appended to the list of headers.

Related tests are also updated.

* Addressed comments.
Following major changes have been made:
- Moved generation of ETag from SQL to backend. Now, to keep consistency in mind, ETag is calculated based on the result of instances to fetch.
- If-None-Match header field is now extracted using the ModelBinder.

* Moved metadata result to a separation ObjectResult custom class to keep the controller clean. This new file checks if the cache is valid and returns 304 (not modified) response, else returns 200 (ok) status.

* Added tests for ETagGenerator.

* Addressed comments.

* Removed unused using statements.

* Updated summary of a method to trigger build of the pull request.