Fixed performance issue with ETag.
Description:
Issue was actually caused by response body not being null. Whenever the response HTTP status code is 304 (not modified), body is expected to be null. If the body is not null, it is translated into an exception which is not reflected in the response, however, that does slow down the response. Once this issue was fixed, performance was improved significantly.

Tests have also been updated to test that the body is null.