Bugfix: ETag Performance Issue. (#265)
* - Added health check feature.
- Uri: /health/check
- Checks for health of Blobs and Sql Server. Returns unhealthy if either of these services is unhealthy.
- Sample output:
    {"overallStatus":"Healthy","details":[{"name":"DicomBlobHealthCheck","status":"Healthy","description":"Successfully connected to the blob data store."},{"name":"SqlServerHealthCheck","status":"Healthy","description":"Successfully connected to the data store."}]}
- Added tests for them as well.

* Updating NSubstitute.dll to the latest version.

* Addressed comments.

* Updated DICOM code after health check shared code was moved to healthcare shared components project.

* Fixed interfaces for the shared health check code.

* Fixed performance issue with ETag.

Description:
Issue was actually caused by response body not being null. Whenever the response HTTP status code is 304 (not modified), body is expected to be null. If the body is not null, it is translated into an exception which is not reflected in the response, however, that does slow down the response. Once this issue was fixed, performance was improved significantly.

Tests have also been updated to test that the body is null.

* Removed unnecessary files and also added the changes in the test files.