Revert STOW Refactor to Fix Performance Regression (#1060)
The number of potential deadlocks has increased significantly with the new BeginAddInstance and EndAddInstance methods. As such, this change reverts the previous refactor to account for potentially changing extended query tags between the start and end. A follow-up PR will update the new implementations in a future SQL schema version