Configure Blob Format for Functions (#1931)
Now that Azure Functions may read blob data for Export, it also needs to know which blob format to use. Furthermore, re-indexing used to simply log whether it could not find a metadata blob because a missing blob meant the entry was deleted sometime between the beginning of re-indexing and now. However, this assumption is incorrect once we introduced a new blob format. Erroneously, the re-index orchestration is ignoring instances whose metadata cannot be found thinking it is because they were deleted when in actuality it is because the blob path changed.

- Add blob migration configuration to functions project
- Make "New" the default blob format for tests
- Update ErrorHref to export input for quality-of-life. Now the operation status can immediately return the complete set of data
- Use lowercase in export paths
- Refactor re-indexing missing blob handling to double-check the blob was truly deleted