Users/lagomesma/accept header refactor (#2220)
Refactor to support multiple headers. This attempts to simplify the way by which we get a header to use based on if and when a user has requested any acceptable headers.
- we still have to use highest priority where we can
- we still have to except when user didn't request anything acceptable
- we have to set the default syntax when user didn't request it in their header and if it was not mandatory for them to do so

We now return a single value (a new accept header with appropriate transfer syntax set) so we do not have both a return value as well as multiple `out` params to track.