Fix Export E2E Test Cleanup (#2660)
The Export E2E test in our CI pipeline may have problems creating the blob container if the Azure Storage service takes awhile to complete a deletion. This change will instead use a dynamic blob container name such that we do not need to wait on the sometimes asynchronous deletion of containers.