M1a - Private preview - External Store - start storing etag and path (#2681)
## Description
M2 Private preview implementation to start storing file path and etags
in sql so they can later be referenced by other operations.

- Added new dbo.FileProperty table with etag and path linked by
instacekey and watermark to dbo.Instance table and propagating data
through InstanceProperties. On STOW, we will now store the path and etag
when instanceKey is provided.
- Added configurable path to ExternalStore options

### Scope
Out of scope and to be handled in M2 (work item 104097) are Get, Delete
and all other operations aside from Update that deal with a blob file
looking at sql stored path and etag.

Out of scope and will add in follow up pr is partitions in file path

Out of scope until further notice is the update operation.

### Additional details
This PR introduces storing file path and etag on STOW when external
store is enabled. Because file path is today a preconfigured path, GET,
DELETE and other operations will continue to use this preconfigured path
until they are updated to instead use what is stored in SQL.

- if DICOM is updated with a new binary taking in the change in this PR,
but the schema has not been upgraded yet, clients will receive an error
asking them to upgrade the schema. This will be true whether external
store is enabled or not.

- if DICOM is is running on an old binary, but the schema has been
upgraded, STOW will not store path and etag in SQL as the instanceKey
and path will be null when passed into the sproc. In this scenario, it
won't be possible for the external store to be enabled as the binary
first needs to be updated, so this is ok. All STOWs will be for an
internal store in this case.

In subsequent PRs, when other operations are changed to look at SQL for
path, if using an internal store, we will continue to use the generated
path. If using an external store we expect all existing indexes will
have also stored a path and etag.

## Related issues
Addresses
[[AB#103363](https://microsofthealth.visualstudio.com/f8da5110-49b1-4e9f-9022-2f58b6124ff9/_workitems/edit/103363)].

## Testing
Programmatic tests as well as manual testing with server

---------

Co-authored-by: Will Sugarman <will.sugarman@microsoft.com>