Add FilePath to ChangeFeed (#2840)
## Description
Add FilePath directly on change feed to avoid costly hash matches when
retrieving. Do this only for created action.
Update action will be addressed in a separate pr.
Fix deleting FileProperty to use instanceKey and index.

## Related issues
Addresses
[[AB#104630](https://microsofthealth.visualstudio.com/f8da5110-49b1-4e9f-9022-2f58b6124ff9/_workitems/edit/104630)].

## Testing
Existing tests modified to ensure data is captured and able to be
retrieved.