get partition name on retrieving deleted instances (#2862)
## Description
Get partition name on retrieving deleted instances to use to generate
file path for external store.

For any binaries using an old schema version, the name will default and
it is not used for file path. Only users with external store are
affected by this.

Work is being done here to have a tighter contract on guaranteeing that
both key and name are non-nulls anytime a partition is present:
https://github.com/microsoft/dicom-server/pull/2807

## Related issues
Addresses
[[AB#105059](https://microsofthealth.visualstudio.com/f8da5110-49b1-4e9f-9022-2f58b6124ff9/_workitems/edit/105059)].

## Testing
Added unit test to ensure nondefault partition used when provided.