Move Delete Metrics to Worker from Health Check (#3183)
The background service health check does not provide additional insight. If SQL or Azure Storage is down, their corresponding health check will fail. And the background service itself will not because it catches all exceptions. If it did somehow crash then the service would crash too.

This PR moves the emission of delete-related metrics to the background service itself.